// sportdash-backup-2025-11-23/_backup/lib-1757716431/types.ts
export type GroupState = "Groen" | "Geel" | "Oranje" | "Rood";

export type Note = {
  id: string;
  text: string;
  author?: string;
  createdAt: string; // ISO string
  starred?: boolean;
};

export type Group = {
  id: string;
  name: string;
  state: GroupState;
  note?: string; // laatste notitie tekst (samenvatting)
  notes?: Note[]; // volledige notitielijst
};